refactor: create schema module
